# swagger.yaml
openapi: 3.0.0
info:
  title: Workout Management API
  description: A complete RESTful API backend for a gym workout tracking application built with Node.js, Express, and PostgreSQL.
  version: 1.0.0
  contact:
    email: support@workoutapi.com
servers:
  - url: http://localhost:3000
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Exercise:
      type: object
      properties:
        id:
          type: integer
          example: 1
        id_user:
          type: integer
          example: 1
        name:
          type: string
          example: "Bench Press"
        description:
          type: string
          nullable: true
          example: "Chest exercise"
        image_url:
          type: string
          nullable: true
          example: "/images/exercises/bench-press.jpg"
        has_pulley:
          type: boolean
          default: false
        is_unilateral:
          type: boolean
          default: false
        bar_weight:
          type: integer
          nullable: true
          enum: [6, 8, 10, 12, 20]
          example: 20
        muscle_group:
          type: string
          example: "Chest"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    WorkoutExercise:
      type: object
      properties:
        id:
          type: integer
        id_exercise:
          type: integer
        sets:
          type: integer
          example: 3
        number:
          type: integer
          example: 1
        name:
          type: string
          example: "Bench Press"
        muscle_group:
          type: string
          example: "Chest"
...
    SessionExercise:
      type: object
      properties:
        id:
          type: integer
        id_exercise:
          type: integer
        number:
          type: integer
          example: 1
        observations:
          type: string
          nullable: true
          example: "Felt heavy today"
        name:
          type: string
          example: "Bench Press"
        muscle_group:
          type: string
          example: "Chest"
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Error message description"
paths:
...
  /api/exercises:
    get:
      tags:
        - Exercises
      summary: Get all user exercises
      security:
        - BearerAuth: []
      parameters:
        - name: muscle_group
          in: query
          description: Filter by muscle group
          schema:
            type: string
            example: "Chest"
      responses:
        '200':
          description: Exercises retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  exercises:
                    type: array
                    items:
                      $ref: '#/components/schemas/Exercise'
    post:
      tags:
        - Exercises
      summary: Create new exercise
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - name
                - muscle_group
              properties:
                name:
                  type: string
                  example: "Bench Press"
                description:
                  type: string
                  example: "Chest exercise"
                muscle_group:
                  type: string
                  example: "Chest"
                has_pulley:
                  type: boolean
                  default: false
                is_unilateral:
                  type: boolean
                  default: false
                bar_weight:
                  type: integer
                  enum: [6, 8, 10, 12, 20]
                image:
                  type: string
                  format: binary
      responses:
        '201':
          description: Exercise created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Exercise created successfully"
                  exercise:
                    $ref: '#/components/schemas/Exercise'
  /api/exercises/{id}:
    get:
      tags:
        - Exercises
      summary: Get exercise by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Exercise retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  exercise:
                    $ref: '#/components/schemas/Exercise'
        '404':
          description: Exercise not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Exercises
      summary: Update exercise
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                muscle_group:
                  type: string
                has_pulley:
                  type: boolean
                is_unilateral:
                  type: boolean
                bar_weight:
                  type: integer
                  enum: [6, 8, 10, 12, 20]
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: Exercise updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Exercise updated successfully"
                  exercise:
                    $ref: '#/components/schemas/Exercise'
    delete:
      tags:
        - Exercises
      summary: Delete exercise
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Exercise deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Exercise deleted successfully"
  /api/exercises/muscle-groups/list:
    get:
      tags:
        - Exercises
      summary: Get all muscle groups
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Muscle groups retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  muscle_groups:
                    type: array
                    items:
                      type: string
                    example: ["Chest", "Back", "Shoulders", "Legs"]
...
  /api/workouts/{id}/exercises:
    post:
      tags:
        - Workouts
      summary: Add exercise to workout
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - exercise_id
                - sets
                - number
              properties:
                exercise_id:
                  type: integer
                  example: 1
                sets:
                  type: integer
                  example: 3
                number:
                  type: integer
                  example: 1
      responses:
        '201':
          description: Exercise added to workout successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Exercise added to workout successfully"
                  workout_exercise:
                    $ref: '#/components/schemas/WorkoutExercise'
  /api/workouts/{workoutId}/exercises/{exerciseId}:
    delete:
      tags:
        - Workouts
      summary: Remove exercise from workout
      security:
        - BearerAuth: []
      parameters:
        - name: workoutId
          in: path
          required: true
          schema:
            type: integer
            example: 1
        - name: exerciseId
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Exercise removed from workout successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Exercise removed from workout successfully"
...
  /api/workout-sessions/{id}/exercises:
    post:
      tags:
        - Workout Sessions
      summary: Add exercise to session
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - exercise_id
                - number
              properties:
                exercise_id:
                  type: integer
                  example: 1
                number:
                  type: integer
                  example: 1
                observations:
                  type: string
                  example: "Felt heavy today"
      responses:
        '201':
          description: Exercise added to workout session successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Exercise added to workout session successfully"
                  session_exercise:
                    $ref: '#/components/schemas/SessionExercise'
...